# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "LMS Authentication API",
    "version": "1.0.0",
    "description": "Learning Management System Authentication API with comprehensive user management features",
    "contact": {
      "name": "API Support",
      "email": "support@lms.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Request success status"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          }
        }
      },
...
      "Note": {
        "type": "object",
        "properties": {
          "noteId": {
            "type": "integer",
            "description": "Note ID"
          },
          "title": {
            "type": "string",
            "description": "Note title"
          },
          "content": {
            "type": "string",
            "description": "Note content"
          },
          "noteType": {
            "type": "string",
            "enum": [
              "text",
              "document",
              "link",
              "multimedia"
            ],
            "description": "Type of note"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Note tags"
          },
          "userId": {
            "type": "integer",
            "description": "Owner user ID"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "required": [
          "title"
        ],
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Note title"
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "description": "Note content"
          },
          "noteType": {
            "type": "string",
            "enum": [
              "text",
              "document",
              "link",
              "multimedia"
            ],
            "default": "text",
            "description": "Type of note"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Note tags"
          },
          "relatedModuleId": {
            "type": "integer",
            "description": "Related learning module ID"
          }
        }
      },
      "UpdateNoteRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255,
            "description": "Note title"
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "description": "Note content"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Note tags"
          },
          "relatedModuleId": {
            "type": "integer",
            "description": "Related learning module ID"
          }
        }
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {}
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/groups/{id}/notes": {
      "get": {
        "summary": "Get notes shared with the group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of notes per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term for note title/content"
          },
          {
            "in": "query",
            "name": "noteType",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "document",
                "link",
                "multimedia"
              ]
            },
            "description": "Filter by note type"
          }
        ],
        "responses": {
          "200": {
            "description": "Group notes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Note"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not a member of this group"
          },
          "404": {
            "description": "Group not found"
          }
        }
      }
    },
    "/notes": {
      "post": {
        "summary": "Create a new note",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              },
              "example": {
                "title": "JavaScript Fundamentals",
                "content": "Variables, functions, and scope concepts",
                "noteType": "text",
                "tags": [
                  "javascript",
                  "programming",
                  "fundamentals"
                ],
                "relatedModuleId": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "note": {
                          "$ref": "#/components/schemas/Note"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/notes/{id}": {
      "get": {
        "summary": "Get a specific note",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Note ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Note retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "note": {
                          "$ref": "#/components/schemas/Note"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "put": {
        "summary": "Update a note",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Note ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteRequest"
              },
              "example": {
                "title": "Advanced JavaScript Concepts",
                "content": "Closures, prototypes, and async programming",
                "tags": [
                  "javascript",
                  "advanced",
                  "async"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "note": {
                          "$ref": "#/components/schemas/Note"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to update this note"
          },
          "404": {
            "description": "Note not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a note",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Note ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Note deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to delete this note"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    },
    "/notes/user/{userId}": {
      "get": {
        "summary": "Get user's notes with pagination and filtering",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "User ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of notes per page"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search term for title/content"
          },
          {
            "in": "query",
            "name": "noteType",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "document",
                "link",
                "multimedia"
              ]
            },
            "description": "Filter by note type"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated tags to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "Notes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Note"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to view these notes"
          }
        }
      }
    },
    "/notes/{id}/share/{groupId}": {
      "post": {
        "summary": "Share a note with a study group",
        "tags": [
          "Notes"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Note ID"
          },
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Group ID to share with"
          }
        ],
        "responses": {
          "200": {
            "description": "Note shared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not authorized to share this note or access this group"
          },
          "404": {
            "description": "Note or group not found"
          }
        }
      }
    },
    "/notes/public": {
      "get": {
        "summary": "Get public notes",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "noteType",
            "schema": {
              "type": "string",
              "enum": [
                "text",
                "image",
                "audio",
                "video",
                "pdf"
              ]
            }
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "oldest",
                "popular"
              ],
              "default": "newest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public notes retrieved successfully"
          }
        }
      }
    },
    "/notes/{noteId}/visibility": {
      "patch": {
        "summary": "Update note visibility",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "noteId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "visibilityScope"
                ],
                "properties": {
                  "visibilityScope": {
                    "type": "string",
                    "enum": [
                      "private",
                      "group",
                      "public"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note visibility updated successfully"
          },
          "403": {
            "description": "Not authorized to update this note"
          },
          "404": {
            "description": "Note not found"
          }
        }
      }
    }
  },
...
}